pyproject.toml
[build-system]
requires = ["flit_core>=3.2", "DiscoverVersion"]
build-backend = "flit_core.buildapi"

[project]
name = 'ce-ui'
description = 'The user-facing interface of contact.engineering.'
requires-python = '>=3.9.0'
dynamic = ['version']
readme = 'README.rst'
license = { file = 'LICENSE' }
authors = [
    { name = "Michael Röttger", email = "info@michael-roettger.de" },
    { name = "Lars Pastewka", email = "lars.pastewka@imtek.uni-freiburg.de" },
]
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Environment :: Web Environment',
    'Framework :: Django',
    'Framework :: Django :: 3.2',
    'Framework :: Celery',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: POSIX',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Programming Language :: Python :: 3.12',
    'Topic :: Internet :: WWW/HTTP',
    'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
]
dependencies = [
    'DiscoverVersion',
    'django-termsandconditions>=2.0.8',
    'celery[redis]',
    'boto3==1.35.99'
]

[project.urls]
homepage = 'https://contact.engineering'
repository = 'https://github.com/ContactEngineering/ce-ui'

[project.entry-points.'topobank.plugins']
ce_ui = 'ce_ui.apps.CEUIPluginConfig'

[project.entry-points.'topobank.third_party_apps']
termsandconditions = 'termsandconditions'

[project.entry-points.'topobank.middleware']
termsandconditions = 'termsandconditions.middleware.TermsAndConditionsRedirectMiddleware'

[project.entry-points.'topobank.context_processors']
fixed_tabs = 'ce_ui.context_processors.fixed_tabs_processor'
